+++ src/main/java/edu/ucar/unidata/rosetta/converters/known/TagUniversalFileFormat.java
@@ -3,7 +3,12 @@
  * See LICENSE for license information.
  */
 
-package edu.ucar.unidata.rosetta.converters;
+/*
+ * Copyright (c) 2012-2018 University Corporation for Atmospheric Research/Unidata
+ * See LICENSE for license information.
+ */
+
+package edu.ucar.unidata.rosetta.converters.known;
 
 import java.io.BufferedReader;
 import java.io.FileInputStream;
